add stub glossary page under data management
